pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "insee-sirene-monitor"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyarrow",
    "requests",
    "python-dotenv",
    "dbt-databricks",
    "databricks-sdk"
]

[tool.setuptools.packages.find]
include = ["ingestion*", "transform*", "utils*", "tasks*"]